+++ CHANGES.rst
@@ -59,6 +59,7 @@ Changelog
 
 **Fixed**
 
+- #1552 Rejection on registration is neither generating rejection pdf nor email
 - #1550 Fix Uncaught TypeError in combogrid
 - #1542 Fix sporadical errors when contacts do not have a valid email address
 - #1540 Fix flushing CCEmail fields in Sample Add Form